+++ scheduler/start-scheduler.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-java $JAVA_OPTS -Djava.library.path=/usr/lib -jar /tmp/elasticsearch-mesos-scheduler.jar $@
+java $JAVA_OPTS -Djava.library.path=/usr/lib -jar /tmp/elasticsearch-mesos-scheduler.jar "$@"